refactor LoginInfo package 이동
