+++ src/main/java/com/jay/codingdojo/atdd/okr/application/user/UserFacade.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Component;
 import com.jay.codingdojo.atdd.okr.domain.guest.service.GuestService;
 import com.jay.codingdojo.atdd.okr.domain.user.ProviderType;
 import com.jay.codingdojo.atdd.okr.domain.user.User;
-import com.jay.codingdojo.atdd.okr.domain.user.service.LoginInfo;
 import com.jay.codingdojo.atdd.okr.domain.user.service.UserWholeInfo;
 import com.jay.codingdojo.atdd.okr.domain.user.service.UserService;
