+++ src/main/java/com/jay/codingdojo/atdd/okr/interfaces/user/LoginResponse.java
@@ -1,7 +1,7 @@
 package com.jay.codingdojo.atdd.okr.interfaces.user;
 
 import com.jay.codingdojo.atdd.okr.domain.user.ProviderType;
-import com.jay.codingdojo.atdd.okr.domain.user.service.LoginInfo;
+import com.jay.codingdojo.atdd.okr.application.user.LoginInfo;
 
 public record LoginResponse(String guestId, String email, String name, ProviderType providerType, String accessToken,
 							String refreshToken) {
